Made mono-Makefile, Fixed all -Wall errors, Fixed some leaks

+++ src/sk_libc/sk_memcpy.c
@@ -3,7 +3,7 @@
 void	*sk_memcpy(void *to_write, void *to_copy, size_t len)
 {
 	unsigned char	*write, *copy;
-	int				i;
+	int			i;
 
 	write = (unsigned char *) to_write;
 	copy = (unsigned char *) to_copy;
@@ -17,7 +17,7 @@ void	*sk_memcpy(void *to_write, void *to_copy, size_t len)
 		return ((char *)to_write + len);
 	}
 	i = -1;
-	while (++i != len)
+	while (++i != (int)len)
 		write[i] = copy[i];
 	return ((char *)to_write + i);
 }
